VORTEX-5581: Force reconnect after proxy change - simplified logic
