+++ open-sphere-base/core/src/main/java/io/opensphere/core/net/NetworkConfigurationOptionsProvider.java
@@ -147,7 +147,7 @@ public class NetworkConfigurationOptionsProvider extends AbstractPreferencesOpti
             myNetworkConfigurationManager.setProxyExclusions("");
         }
 
-        ((NetworkConfigurationManagerImpl)myNetworkConfigurationManager).notifyNetworkConfigChanged();
+        ((NetworkConfigurationManagerImpl)myNetworkConfigurationManager).notifyChanged();
     }
 
     @Override
